# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Games4King Workplace OS API
  version: 1.0.0
  description: Internal employee portal API for Games4King
  license:
    name: MIT
servers:
  - url: /api
    description: API proxy (via Next.js rewrites)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login with identifier + password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                password:
                  type: string
                device_name:
                  type: string
      responses:
        '200':
          description: Success — returns access token + user
        '422':
          description: Invalid credentials
        '423':
          description: Account locked
  /auth/refresh:
    get:
      tags:
        - Auth
      summary: Refresh access token via HttpOnly refresh cookie
      security: []
      responses:
        '200':
          description: New access token
        '401':
          description: Invalid or expired refresh token
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset link sent
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password with token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - email
                - password
                - password_confirmation
              properties:
                token:
                  type: string
                email:
                  type: string
                password:
                  type: string
                password_confirmation:
                  type: string
      responses:
        '200':
          description: Password reset successfully
...
  /auth/role-select:
    post:
      tags:
        - Auth
      summary: Switch active role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
      responses:
        '200':
          description: Role switched — new access token
...
  /auth/sessions/{id}:
    delete:
      tags:
        - Auth
      summary: Revoke a specific session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Session revoked
...
  /auth/profile:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                avatar_url:
                  type: string
      responses:
        '200':
          description: Updated Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /profile/avatar:
    post:
      tags:
        - Profile
      summary: Upload avatar image
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar URL
  /directory:
    get:
      summary: Get employee directory
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Directory list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
  /directory/{id}:
    get:
      tags:
        - Directory
      summary: View employee details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee details
  /directory/{id}/send-message:
    post:
      tags:
        - Directory
      summary: Send a direct message to an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Message sent
...
  /attendance/me/day/{date}:
    get:
      tags:
        - Attendance
      summary: Get specific day attendance
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Day attendance with events
  /attendance/admin/overview:
    get:
      tags:
        - Attendance
      summary: Admin overview of all attendance
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: department_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated attendance overview
...
  /attendance/hr/graph:
    get:
      tags:
        - Attendance
      summary: HR attendance graph data
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum:
              - weekly
              - monthly
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Graph data
...
  /leave-requests:
    get:
      tags:
        - Leave
      summary: List own leave requests
      responses:
        '200':
          description: Paginated leave requests
    post:
      tags:
        - Leave
      summary: Submit a new leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - start_date
                - end_date
                - type
                - reason
              properties:
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                type:
                  type: string
                  enum:
                    - casual
                    - sick
                    - earned
                    - unpaid
                reason:
                  type: string
      responses:
        '201':
          description: Leave request created
  /leave-requests/history:
    get:
      tags:
        - Leave
      summary: Leave history with filters
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated leave history
  /leave-requests/{id}:
    get:
      tags:
        - Leave
      summary: View a specific leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Leave request details
...
  /approvals/{id}/decision:
    post:
      tags:
        - Leave
      summary: Approve or reject a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
              properties:
                decision:
                  type: string
                  enum:
                    - approved
                    - rejected
                reason:
                  type: string
      responses:
        '200':
          description: Decision recorded
  /holidays:
    get:
      tags:
        - Leave
      summary: List holidays for a year
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of holidays
...
  /notifications/{id}/mark-read:
    post:
      tags:
        - Notifications
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Marked as read
...
  /users/{user}:
    get:
      tags:
        - Admin - Users
      summary: Show user details
      parameters:
        - name: user
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User object
    put:
      tags:
        - Admin - Users
      summary: Update user
      parameters:
        - name: user
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated user
    delete:
      tags:
        - Admin - Users
      summary: Delete user
      parameters:
        - name: user
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
  /users/{id}/reset-password:
    post:
      tags:
        - Admin - Users
      summary: Reset a user's password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Password reset
...
  /companies/{company}:
    get:
      tags:
        - Admin - Companies
      summary: Show company
      parameters:
        - name: company
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Company object
    put:
      tags:
        - Admin - Companies
      summary: Update company
      parameters:
        - name: company
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated company
    delete:
      tags:
        - Admin - Companies
      summary: Delete company
      parameters:
        - name: company
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Company deleted
...
  /departments/{department}:
    get:
      tags:
        - Admin - Departments
      summary: Show department
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Department object
    put:
      tags:
        - Admin - Departments
      summary: Update department
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated department
    delete:
      tags:
        - Admin - Departments
      summary: Delete department
      parameters:
        - name: department
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Department deleted
...
  /designations/{designation}:
    get:
      tags:
        - Admin - Designations
      summary: Show designation
      parameters:
        - name: designation
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Designation object
    put:
      tags:
        - Admin - Designations
      summary: Update designation
      parameters:
        - name: designation
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated designation
    delete:
      tags:
        - Admin - Designations
      summary: Delete designation
      parameters:
        - name: designation
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Designation deleted
...
  /work-schedules/{id}:
    put:
      tags:
        - Settings
      summary: Update work schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
...
  /projects/{project}:
    get:
      tags:
        - Projects
      summary: Show project
      parameters:
        - name: project
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project object
    put:
      tags:
        - Projects
      summary: Update project
      parameters:
        - name: project
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated project
    delete:
      tags:
        - Projects
      summary: Delete project
      parameters:
        - name: project
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project deleted
...
  /tasks/{task}:
    get:
      tags:
        - Tasks
      summary: Show task
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task object
    put:
      tags:
        - Tasks
      summary: Update task
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated task
    delete:
      tags:
        - Tasks
      summary: Delete task
      parameters:
        - name: task
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted
  /tasks/{id}/submit-review:
    post:
      tags:
        - Tasks
      summary: Submit task for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Submitted
  /tasks/{id}/comments:
    post:
      tags:
        - Tasks
      summary: Add comment to task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comment added
...
  /qa-forms/{id}:
    get:
      tags:
        - QA
      summary: Show QA form
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: QA form object
...
  /saved-views/{id}:
    delete:
      tags:
        - Views
      summary: Delete a saved view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: View deleted
...
  /conversations/{id}/messages:
    get:
      tags:
        - Chat
      summary: List messages in conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Messages
    post:
      tags:
        - Chat
      summary: Send message in conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Message sent
...
  /quick-notes/{id}:
    delete:
      tags:
        - Communication
      summary: Delete quick note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Note deleted
...